assign continents on pre-made maps as well
