+++ core/src/com/unciv/logic/civilization/diplomacy/DiplomacyManager.kt
@@ -352,7 +352,6 @@ class DiplomacyManager() {
             return true
         return hasOpenBorders
     }
-
     //endregion
 
     //region state-changing functions
